refactor: quantization for moe layers
