+++ python/sglang/srt/layers/moe/fused_moe_triton/__init__.py
@@ -1,16 +1,16 @@
 from contextlib import contextmanager
 from typing import Any, Dict, Optional
 
-import sglang.srt.layers.moe.fused_moe_triton.fused_moe  # noqa
 from sglang.srt.layers.moe.fused_moe_triton.fused_moe import (
     fused_experts,
     get_config_file_name,
 )
 from sglang.srt.layers.moe.fused_moe_triton.layer import (
     FusedMoE,
-    FusedMoEMethodBase,
     FusedMoeWeightScaleSupported,
 )
+
+FusedMoEMethodBase = None
 
 _config: Optional[Dict[str, Any]] = None
